Add slightly smarter content-type detection via `mime`.
Ideally we would be using something like `libmagic` to do this, but it's kind of a heavyweight solution and platform dependant. So for now, just use the handy `mime` module to lookup a best guess based on the extension. If someone needs something more accurate they can always set `contentType`.