Add missing `isNull` check for mapping vinyl metadata.
When requesting only metadata (i.e. `{ read: false }`) the check for the `.length` property on the file's contents will fail because the file has no contents. A check is added to deal with this case and a corresponding test has been provided.

Additionally the README has been updated to include an otherwise missing module, and the examples folder now includes an example for just retrieving the metadata.

This fixes https://github.com/izaakschroeder/vinyl-s3/issues/9.